pyproject.toml
[tool.poetry]
name = "career-tracker-hr-api"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "career_tracker_hr_api"}]

[tool.poetry.dependencies]
python = "^3.10"
django = "4.2.6"
python-dotenv = "^1.0.0"
psycopg2-binary = "^2.9.9"
pillow = "^10.1.0"
djangorestframework = "^3.14.0"


[tool.poetry.group.dev.dependencies]
pre-commit = "^3.5.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
